[FE-215] refactor: common폴더 변경으로 인한 import 수정
